fix(player): working embedded playback, controls, and fullscreen
Bring-up fixes for the mpv player: use the 2-arg loadfile form (mpv rejected the "replace" flag with error -4), poll mpv ~5x/sec to push PlayerState to the UI (embedded property-change events were unreliable), raise the mpv child window above the WebView (z-order) so video renders instead of showing black, unpause on file-loaded, and convert the video-pane rect using the window scale factor. Adds true display fullscreen (allow-set-fullscreen) and a dbcheck example.

+++ src-tauri/src/commands/player.rs
@@ -83,10 +83,10 @@ pub fn player_prev(state: State<AppState>) -> Result<()> {
 #[tauri::command]
 pub fn player_set_rect(
     state: State<AppState>,
-    x: i32,
-    y: i32,
-    w: i32,
-    h: i32,
+    x: f64,
+    y: f64,
+    w: f64,
+    h: f64,
 ) -> Result<()> {
     existing(&state, |p| p.set_rect(x, y, w, h))
 }
